# Perf Testing and Improvement - Beast Ingest Allocations