--- a/lib/producer/sbs1.go
+++ b/lib/producer/sbs1.go
@@ -5,7 +5,7 @@ import (
 	"plane.watch/lib/tracker/sbs1"
 )
 
-func (p *producer) sbsScanner(scan *bufio.Scanner) error {
+func (p *Producer) sbsScanner(scan *bufio.Scanner) error {
 	for scan.Scan() {
 		line := scan.Text()
 		p.addFrame(sbs1.NewFrame(scan.Text()), &p.FrameSource)