Change the beast scanner to not do a memory allocation for every message it processes, but rather use a common scratch []byte and return out of it

Testing with

  go test -bench . -count 5 -benchmem

We go from an original

```
goos: linux
goarch: amd64
pkg: plane.watch/lib/producer
cpu: AMD Ryzen 7 3800X 8-Core Processor
BenchmarkScanBeast-16                 25          42657336 ns/op         5332599 B/op     272665 allocs/op
BenchmarkScanBeast-16                 27          43686231 ns/op         5332491 B/op     272665 allocs/op
BenchmarkScanBeast-16                 31          44063729 ns/op         5332415 B/op     272665 allocs/op
BenchmarkScanBeast-16                 27          41658114 ns/op         5332416 B/op     272665 allocs/op
BenchmarkScanBeast-16                 31          40061668 ns/op         5332418 B/op     272665 allocs/op
PASS
ok      plane.watch/lib/producer        6.489s
```
down to
```
goos: linux
goarch: amd64
pkg: plane.watch/lib/producer
cpu: AMD Ryzen 7 3800X 8-Core Processor
BenchmarkScanBeast-16                 74          16236748 ns/op            4177 B/op          3 allocs/op
BenchmarkScanBeast-16                 75          16548997 ns/op            4177 B/op          3 allocs/op
BenchmarkScanBeast-16                 69          16108247 ns/op            4177 B/op          3 allocs/op
BenchmarkScanBeast-16                 68          15988906 ns/op            4177 B/op          3 allocs/op
BenchmarkScanBeast-16                 69          16162476 ns/op            4177 B/op          3 allocs/op
PASS
ok      plane.watch/lib/producer        7.363s
```