@shaneshort 2022-10-15
new fone who dis?